+++ core-ng/src/main/java/core/framework/api/search/SearchException.java
@@ -6,6 +6,10 @@ package core.framework.api.search;
 public class SearchException extends RuntimeException {
     private static final long serialVersionUID = -1161076239987892452L;
 
+    public SearchException(String message) {
+        super(message);
+    }
+
     public SearchException(Throwable cause) {
         super(cause);
     }